let examples have a chance to exit normally

+++ examples/test_examples.py
@@ -219,10 +219,9 @@ try:
     import sys
     print("test complete")
     sys.stdout.flush()
-    import time
-    while True:  ## run a little event loop
-        pg.QtGui.QApplication.processEvents()
-        time.sleep(0.01)
+    pg.Qt.QtCore.QTimer.singleShot(1000, pg.Qt.QtWidgets.QApplication.quit)
+    pg.Qt.QtWidgets.QApplication.instance().exec_()
+    # pg.exit()
 except:
     print("test failed")
     raise
@@ -259,8 +258,13 @@ except:
         if output.endswith('test failed'):
             fail = True
             break
-    time.sleep(1)
-    process.kill()
+    start = time.time()
+    killed = False
+    while process.poll() is None:
+        time.sleep(0.1)
+        if time.time() - start > 2.0 and not killed:
+            process.kill()
+            killed = True
     #res = process.communicate()
     res = (process.stdout.read(), process.stderr.read())
     if (fail or
